+++ MetovaTestKitTests/UIKitTestingTests/CellTestingTests/MTKTableViewTestCellTests.swift
@@ -41,7 +41,7 @@ class TableViewTestCellTests: MTKBaseTestCase {
         tableView.dataSource = dataSource
         
         let testExecutedExpectation = expectation(description: "Test block should be executed!")
-        MTKTestCell(in: tableView, at: indexPath, as: MockTableCellTypeA.self) { testCell in
+        MTKTestCell(in: tableView, at: indexPath, as: MockTableCellTypeA.self) { _ in
             testExecutedExpectation.fulfill()
         }
         wait(for: [testExecutedExpectation], timeout: 0)
@@ -52,8 +52,8 @@ class TableViewTestCellTests: MTKBaseTestCase {
         let indexPath = IndexPath(row: 0, section: 0)
         
         let description = "failed - \(tableView) does not have a dataSource"
-        expectTestFailure(TestFailureExpectation(description: description, lineNumber: #line+1)) {
-            MTKTestCell(in: tableView, at: indexPath, as: UITableViewCell.self) { testCell in
+        expectTestFailure(BasicTestFailureExpectation(description: description, lineNumber: #line+1)) {
+            MTKTestCell(in: tableView, at: indexPath, as: UITableViewCell.self) { _ in
                 XCTFail("Test closure should not be called when cell can not be properly instantiated and cast.")
             }
         }
@@ -67,8 +67,8 @@ class TableViewTestCellTests: MTKBaseTestCase {
         tableView.dataSource = dataSource
         
         let description = "failed - \(tableView) does not have a cell at \(indexPath)"
-        expectTestFailure(TestFailureExpectation(description: description, lineNumber: #line+1)) {
-            MTKTestCell(in: tableView, at: indexPath, as: UITableViewCell.self) { testCell in
+        expectTestFailure(BasicTestFailureExpectation(description: description, lineNumber: #line+1)) {
+            MTKTestCell(in: tableView, at: indexPath, as: UITableViewCell.self) { _ in
                 XCTFail("Test closure should not be called when cell can not be properly instantiated and cast.")
             }
         }
@@ -82,8 +82,8 @@ class TableViewTestCellTests: MTKBaseTestCase {
         tableView.dataSource = dataSource
         
         let description = "failed - Cell at \(indexPath) expected to be \(MockTableCellTypeA.self) was actually \(MockTableCellTypeB.self)"
-        expectTestFailure(TestFailureExpectation(description: description, lineNumber: #line+1)) {
-            MTKTestCell(in: tableView, at: indexPath, as: MockTableCellTypeA.self) { testCell in
+        expectTestFailure(BasicTestFailureExpectation(description: description, lineNumber: #line+1)) {
+            MTKTestCell(in: tableView, at: indexPath, as: MockTableCellTypeA.self) { _ in
                 XCTFail("Test closure should not be called when cell can not be properly instantiated and cast.")
             }
         }